Adds validation and testing to job posting form
In the JobPosting model, there is now
a check that validates the presence of
data in each of the fields.

In the test folder, new tests check and
verify the condition of a whole fully
created form. Additionally, the tests
check to make sure all the fields are
filled with data and are valid data.