Adds ability for admin to edit and approve employer job postings
Adds checkbox to form page for admins to approve jobs. Checked
check boxes will indicate an approved job posting that will be
displayed on a separate page.

When a non-administrative user is on the website they
will not have the ability to approve or disapprove jobs.
They will also not have the ability to edit job postings. Their
privilege level is set to read-only.