Adds search bar and ability to search all database fields
Removes the ability for admin to register, a new seeded admin is in
the database.

Creates a new partial file for the job_postings. View pages that use
the list of job postings will be using the partial to render the view.

Adds admin login to the application.html.erb so it will be globally
available in all of the web pages.

Routes user immediately to the index page instead of the display
page, which is what it was previously doing.