Adds working unit tests that use the fixtures from job_postings.yml
One test is added for each of the form fields to test that the data
entered should be filled in.

One test is added for the search bar functionality per each field in
the database with the exception of,  job_type, job_type_id, time_
created, time_updated, and approved. The job_type fields should
be placed in a separate test file.