Adds employer dashboard for each approved user
adds index.html.erb to the user folder

adds invisible form field to the form to input the current_user.id for
each submission.

After submitting, the newly submitted job will appear on that employers
dashboard page along with a column letting the user know whether or not
the job has been approved by an admin.